requirements.txt
# Django and Related Packages
Django==5.0.2
asgiref==3.7.2

# Django REST Framework and Authentication
djangorestframework==3.15.2
djangorestframework-api-key==3.0.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
cffi==1.16.0
pycparser==2.21

# AWS and Storage
django-storages==1.14.2
boto3==1.34.54
botocore==1.34.54
s3transfer==0.10.0
jmespath==1.0.1
python-dateutil==2.8.2
six==1.16.0
urllib3>=1.25.4,<2.1
certifi==2024.2.2

# osu! Related Packages
osrparse==6.0.2
ossapi==3.4.4
rosu-pp-py==3.1.0

# HTTP and Networking
requests==2.31.0
httpx==0.27.0
anyio==4.3.0
sniffio==1.3.1
httpcore==1.0.4
h11==0.14.0
idna==3.6
charset-normalizer==3.3.2

# Authentication and Security
oauthlib==3.2.2
requests-oauthlib==1.3.1

# Environment Variables and Configuration
python-dotenv==1.0.1

# Logging
python-json-logger==2.0.7

# Date and Time Handling
arrow==1.3.0
tzdata==2023.4
isoduration==20.11.0
python-dateutil==2.8.2

# Utilities and Miscellaneous
colorama==0.4.6
decorator==5.1.1
attrs==23.2.0
overrides==7.7.0
packaging==23.2
platformdirs==4.2.0
PyYAML==6.0.1
typing-utils==0.1.0
nltk==3.8.1
better_profanity==0.7.0

# Data Parsing and Validation
jsonschema==4.21.1
fastjsonschema==2.19.1
jsonschema-specifications==2023.12.1
jsonpointer==2.4
referencing==0.33.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
webcolors==1.13

# HTML and XML Processing
beautifulsoup4==4.12.3
soupsieve==2.5
bleach==6.1.0
defusedxml==0.7.1
MarkupSafe==2.1.5
Jinja2==3.1.3

# Development and Debugging Tools (Optional)
debugpy==1.8.1
asttokens==2.4.1
executing==2.0.1
pure-eval==0.2.2
stack-data==0.6.3
jedi==0.19.1
prompt-toolkit==3.0.43
traitlets==5.14.1
tornado==6.4

# Build and Packaging Tools (Optional)
setuptools==69.1.1
wheel==0.42.0

# WSGI Server
gunicorn==21.2.0

# Windows-Specific Packages
pywin32==306; sys_platform == 'win32'
pywin32-ctypes==0.2.2; sys_platform == 'win32'
pywinpty==2.0.13; sys_platform == 'win32'
pefile==2023.2.7; sys_platform == 'win32'
